Purchases: Move CreditCards section to /me/purchases folder
