+++ assets/stylesheets/_components.scss
@@ -176,8 +176,6 @@
 @import 'me/connected-application-item/style';
 @import 'me/connected-applications/style';
 @import 'me/connected-application-icon/style';
-@import 'me/credit-cards/credit-card-delete';
-@import 'me/credit-cards/credit-cards';
 @import 'me/get-apps/style';
 @import 'me/happychat/style';
 @import 'me/help/help-contact-confirmation/style';
@@ -206,6 +204,8 @@
 @import 'me/purchases/cancel-purchase/style';
 @import 'me/purchases/components/loading-placeholder/style';
 @import 'me/purchases/confirm-cancel-domain/style';
+@import 'me/purchases/credit-cards/credit-card-delete';
+@import 'me/purchases/credit-cards/credit-cards';
 @import 'me/purchases/list/item/style';
 @import 'me/purchases/list/site/style';
 @import 'me/purchases/manage-purchase/style';
