+++ client/me/billing-history/main.jsx
@@ -11,7 +11,7 @@ var observe = require( 'lib/mixins/data-observe' ),
 	Card = require( 'components/card' ),
 	MeSidebarNavigation = require( 'me/sidebar-navigation' ),
 	config = require( 'config' ),
-	CreditCards = require( 'me/credit-cards' ),
+	CreditCards = require( 'me/purchases/credit-cards' ),
 	eventRecorder = require( 'me/event-recorder' ),
 	PurchasesHeader = require( '../purchases/list/header' ),
 	BillingHistoryTable = require( './billing-history-table' ),
